+++ src/toolkit.rs
@@ -5,7 +5,7 @@ use crate::{
     internal::generate_internal_table_config,
     validate::{validate_row_tx, validate_rows_constraints, validate_rows_intra},
     valve::{
-        ValveCell, ValveCellMessage, ValveColumnConfig, ValveConfig, ValveConstraintConfig,
+        Valve, ValveCell, ValveCellMessage, ValveColumnConfig, ValveConfig, ValveConstraintConfig,
         ValveDatatypeConfig, ValveError, ValveForeignConstraint, ValveMessage, ValveRow,
         ValveRuleConfig, ValveSpecialConfig, ValveTableConfig, ValveTreeConstraint,
     },
@@ -2098,14 +2098,145 @@ pub fn get_label_for_column(
 
 /// Given a table name, a column name, and a new name for the column, rename the column to the
 /// new name in the database table and associated views, using the given database transaction.
-pub async fn rename_db_column_tx(
-    db_kind: &DbKind,
-    config: &ValveConfig,
+pub async fn rename_column_tx(
+    valve: &Valve,
     tx: &mut Transaction<'_, sqlx::Any>,
     table: &str,
     column: &str,
-    new_col_name: &str,
+    new_name: &str,
+    new_label: &Option<String>,
 ) -> Result<()> {
+    // Update from() structures in the column table that refer to the column.
+    let mut structure_params = vec![];
+    let mut where_params = vec![table];
+    let structure_sql = {
+        let mut sql_lines = vec![];
+        // Look through foreign constraints:
+        for (fkey_table, fkeys) in valve.config.constraint.foreign.iter() {
+            for fkey in fkeys {
+                if fkey.ftable == table && fkey.fcolumn == column {
+                    // Add the table that has a foreign key referencing the renamed column to
+                    // the list of tables that need to be updated:
+                    where_params.push(fkey_table);
+                    sql_lines.push(format!(
+                        r#"WHEN "structure" = {SQL_PARAM} THEN {SQL_PARAM}"#
+                    ));
+                    structure_params.push(fkey.original.to_string());
+                    let new_name = {
+                        if new_name.contains(char::is_whitespace) {
+                            &format!("'{new_name}'")
+                        } else {
+                            new_name
+                        }
+                    };
+                    structure_params
+                        .push(format!("from({ftable}.{new_name})", ftable = fkey.ftable));
+                }
+            }
+        }
+
+        // Update tree() structures that refer to the renamed column:
+        for tkey in valve
+            .config
+            .constraint
+            .tree
+            .get(table)
+            .expect("No tree found for table")
+        {
+            if tkey.child.as_str() == column {
+                sql_lines.push(format!(
+                    r#"WHEN "structure" = {SQL_PARAM} THEN {SQL_PARAM}"#
+                ));
+                structure_params.push(tkey.original.to_string());
+                let new_name = {
+                    if new_name.contains(char::is_whitespace) {
+                        &format!("'{new_name}'")
+                    } else {
+                        new_name
+                    }
+                };
+                let new_cond = format!("tree({new_name})");
+                structure_params.push(new_cond);
+            }
+        }
+
+        // Only actually build the structure clause if there were matching conditions:
+        if !sql_lines.is_empty() {
+            format!(
+                r#""structure" = CASE
+                         {case_clause}
+                         ELSE "structure"
+                       END, "#,
+                case_clause = sql_lines.join("\n")
+            )
+        } else {
+            "".to_string()
+        }
+    };
+
+    let where_sql = format!(
+        r#"WHERE "table" IN ({})"#,
+        where_params
+            .iter()
+            .map(|_| SQL_PARAM)
+            .collect::<Vec<_>>()
+            .join(", "),
+    );
+
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "column"
+                      SET {structure_sql}
+                          "column" = CASE WHEN "column" = {SQL_PARAM} THEN {SQL_PARAM}
+                                          ELSE "column"
+                                     END,
+                          "label" = CASE WHEN "column" = {SQL_PARAM} THEN {SQL_PARAM}
+                                          ELSE "label"
+                                     END
+                      {where_sql}"#
+        ),
+    );
+
+    let mut query = sqlx_query(&sql);
+    for param in &structure_params {
+        query = query.bind(param);
+    }
+    query = query
+        .bind(column)
+        .bind(new_name)
+        .bind(column)
+        .bind(new_label);
+    for param in &where_params {
+        query = query.bind(param);
+    }
+    query.execute(tx.acquire().await?).await?;
+
+    // Now update the when_column and/or the then_column of any rules that refer to the
+    // datatype:
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "rule"
+                     SET "when column" = CASE
+                       WHEN "when column" = {SQL_PARAM} THEN {SQL_PARAM}
+                       ELSE "when column"
+                     END, "then column" = CASE
+                       WHEN "then column" = {SQL_PARAM} THEN {SQL_PARAM}
+                       ELSE "then column"
+                     END
+                   WHERE "table" = {SQL_PARAM}"#
+        ),
+    );
+
+    let query = sqlx_query(&sql)
+        .bind(column)
+        .bind(new_name)
+        .bind(column)
+        .bind(new_name)
+        .bind(table);
+    query.execute(tx.acquire().await?).await?;
+
     // Although PostgreSQL supports an ALTER VIEW statement, SQLite does not, so we have to
     // manually drop and recreate views in that case. For simplicitly, then, we use the manual
     // method in both cases.
@@ -2120,22 +2251,266 @@ pub async fn rename_db_column_tx(
 
     // Next, rename the column in both the table and its corresponding conflict table:
     for suffix in ["", "_conflict"] {
-        let sql = format!(
-            r#"ALTER TABLE "{table}{suffix}" RENAME COLUMN "{column}" TO "{new_col_name}""#
-        );
+        let sql =
+            format!(r#"ALTER TABLE "{table}{suffix}" RENAME COLUMN "{column}" TO "{new_name}""#);
         let query = sqlx_query(&sql);
         query.execute(tx.acquire().await?).await?;
     }
 
     // Now recreate both views:
-    let sql = get_sql_for_standard_view(table, db_kind);
+    let sql = get_sql_for_standard_view(table, &valve.db_kind);
     let query = sqlx_query(&sql);
     query.execute(tx.acquire().await?).await?;
     // For text views, we need to indicate the change to the column name via a HashMap from the
     // old name to the new:
     let mut changes = HashMap::new();
-    changes.insert(column, new_col_name);
-    let sql = get_sql_for_text_view(&config.table, table, None, Some(changes), db_kind);
+    changes.insert(column, new_name);
+    let sql = get_sql_for_text_view(
+        &valve.config.table,
+        table,
+        None,
+        Some(changes),
+        &valve.db_kind,
+    );
+    let query = sqlx_query(&sql);
+    query.execute(tx.acquire().await?).await?;
+
+    Ok(())
+}
+
+/// TODO: Add docstring here
+pub async fn rename_datatype_tx(
+    valve: &Valve,
+    datatype: &str,
+    new_name: &str,
+    tx: &mut Transaction<'_, sqlx::Any>,
+) -> Result<()> {
+    let cols_for_query = valve
+        .config
+        .table
+        .get("datatype")
+        .expect("Datatype table not found in config")
+        .column
+        .keys()
+        .filter(|col| *col != "datatype")
+        .map(|col| format!(r#""{col}""#))
+        .collect::<Vec<_>>();
+
+    let main_column_list = cols_for_query.join(", ");
+    let t2_column_list = cols_for_query
+        .iter()
+        .map(|col| format!("t2.{col}"))
+        .collect::<Vec<_>>()
+        .join(", ");
+
+    // We begin by saving the row number and row order of the datatype to be renamed:
+    let (saved_rn, saved_ro) = {
+        let sql = local_sql_syntax(
+            &valve.db_kind,
+            &format!(
+                r#"SELECT "row_number", "row_order" FROM "datatype"
+                        WHERE "datatype" = {SQL_PARAM}"#
+            ),
+        );
+        let row = sqlx_query(&sql)
+            .bind(datatype)
+            .fetch_one(tx.acquire().await?)
+            .await?;
+        let rn: i64 = row.try_get::<i64, &str>("row_number")?;
+        let ro: i64 = row.try_get::<i64, &str>("row_order")?;
+        (rn, ro)
+    };
+
+    // We then add a new datatype corresponding to the new name:
+    let (rn, ro) = get_next_new_row_tx(&valve.db_kind, tx, "datatype").await?;
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"INSERT INTO "datatype"
+                   ("row_number", "row_order", "datatype", {main_column_list})
+                   SELECT
+                      {rn} AS row_number,
+                      {ro} as row_order,
+                      {SQL_PARAM} as "datatype",
+                      {t2_column_list}
+                     FROM "datatype" t2
+                    WHERE t2."datatype" = {SQL_PARAM}"#
+        ),
+    );
+    let query = sqlx_query(&sql).bind(new_name).bind(datatype);
+    query.execute(tx.acquire().await?).await?;
+
+    // Now update any columns that used the old name as their datatype to the new name:
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "column"
+                     SET "datatype" = CASE
+                       WHEN "datatype" = {SQL_PARAM} THEN {SQL_PARAM}
+                       ELSE "datatype"
+                     END, "nulltype" = CASE
+                       WHEN "nulltype" = {SQL_PARAM} THEN {SQL_PARAM}
+                       ELSE "nulltype"
+                     END"#
+        ),
+    );
+    let query = sqlx_query(&sql)
+        .bind(datatype)
+        .bind(new_name)
+        .bind(datatype)
+        .bind(new_name);
+    query.execute(tx.acquire().await?).await?;
+
+    let maybe_add_to_sql = |column: &str,
+                            ccond: &CompiledCondition,
+                            sql_lines: &mut Vec<String>,
+                            sql_params: &mut Vec<String>| {
+        if let ValueType::List(reference_dt, list_sep) = &ccond.value_type {
+            if reference_dt == datatype {
+                let old_condition = ccond.original.to_string();
+                let new_name = {
+                    if new_name.contains(char::is_whitespace) {
+                        &format!("'{new_name}'")
+                    } else {
+                        new_name
+                    }
+                };
+                let new_condition = format!("list({new_name}, '{list_sep}')");
+                sql_lines.push(format!(r#"WHEN "{column}" = {SQL_PARAM} THEN {SQL_PARAM}"#));
+                sql_params.push(old_condition);
+                sql_params.push(new_condition);
+            }
+        }
+    };
+
+    // Look through the datatype config for any datatypes that are list() types and
+    // that refer to the current datatype as their base type. Then construct a SQL CASE
+    // statement to update the "condition" column of the datatype table appropriately given the
+    // renaming.
+    let mut condition_params = vec![];
+    let condition_sql = {
+        let mut sql_lines = vec![];
+        for (_, dt_condition) in valve.datatype_conditions.iter() {
+            maybe_add_to_sql(
+                "condition",
+                dt_condition,
+                &mut sql_lines,
+                &mut condition_params,
+            );
+        }
+        if !sql_lines.is_empty() {
+            format!(
+                r#""condition" = CASE
+                          {case_clause}
+                          ELSE "condition"
+                        END, "#,
+                case_clause = sql_lines.join("\n")
+            )
+        } else {
+            "".to_string()
+        }
+    };
+
+    // In addition to the (list) conditions of any datatypes that referred to the old name,
+    // any datatypes whose parents refer to the datatype by its old name need to be updated
+    // as well:
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "datatype"
+                    SET {condition_sql}"parent" = CASE
+                      WHEN "parent" = {SQL_PARAM} THEN {SQL_PARAM}
+                      ELSE "parent"
+                    END"#
+        ),
+    );
+    let mut query = sqlx_query(&sql);
+    for condition_param in &condition_params {
+        query = query.bind(condition_param);
+    }
+    for parent_param in [datatype, new_name] {
+        query = query.bind(parent_param);
+    }
+
+    query.execute(tx.acquire().await?).await?;
+
+    // Now update the when_column and/or the then_column of any rules that refer to the
+    // datatype:
+    let mut wcond_params = vec![];
+    let mut tcond_params = vec![];
+    // This closure is used to generate the SET clauses for "when condition" and
+    // "then condition" in the UPDATE statement for the rule table, and to populate
+    // wcond_params and tcond_params:
+    let get_rule_cond_sql = |column: &str, cond_params: &mut Vec<String>| {
+        let mut sql_lines = vec![];
+        // When/then datatype conditions that match the renamed datatype should
+        // be renamed as well:
+        sql_lines.push(format!(r#"WHEN "{column}" = {SQL_PARAM} THEN {SQL_PARAM}"#));
+        cond_params.push(datatype.to_string());
+        cond_params.push(new_name.to_string());
+
+        // When/then datatype conditions that are of the list() type, whose reference
+        // datatype corresponds to the given datatype, should also be renamed:
+        for (_, column_rules) in valve.rule_conditions.iter() {
+            for (_, conditions) in column_rules.iter() {
+                for cond in conditions.iter() {
+                    if column == "when condition" {
+                        maybe_add_to_sql(column, &cond.when, &mut sql_lines, cond_params);
+                    } else if column == "then condition" {
+                        maybe_add_to_sql(column, &cond.then, &mut sql_lines, cond_params);
+                    } else {
+                        panic!("Not a recognized condition column: {column}");
+                    }
+                }
+            }
+        }
+
+        // Only actually build the clause if there were matching conditions:
+        if !sql_lines.is_empty() {
+            format!(
+                r#""{column}" = CASE
+                          {case_clause}
+                          ELSE "{column}"
+                        END"#,
+                case_clause = sql_lines.join("\n")
+            )
+        } else {
+            "".to_string()
+        }
+    };
+    let when_condition_sql = get_rule_cond_sql("when condition", &mut wcond_params);
+    let then_condition_sql = get_rule_cond_sql("then condition", &mut tcond_params);
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "rule"
+                     SET {when_condition_sql}, {then_condition_sql}"#
+        ),
+    );
+    let mut query = sqlx_query(&sql);
+    for param in &wcond_params {
+        query = query.bind(param);
+    }
+    for param in &tcond_params {
+        query = query.bind(param);
+    }
+    query.execute(tx.acquire().await?).await?;
+
+    // Now delete the old datatype name from the database:
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(r#"DELETE FROM "datatype" WHERE "datatype" = {SQL_PARAM}"#),
+    );
+    let query = sqlx_query(&sql).bind(datatype);
+    query.execute(tx.acquire().await?).await?;
+
+    // Finally update the newly added datatype's row number and row order to the values
+    // that we saved earlier, to complete the logical rename operation:
+    let sql = format!(
+        r#"UPDATE "datatype"
+                  SET "row_number" = {saved_rn}, "row_order" = {saved_ro}
+                WHERE "row_number" = {rn}"#
+    );
     let query = sqlx_query(&sql);
     query.execute(tx.acquire().await?).await?;
 
@@ -2144,13 +2519,156 @@ pub async fn rename_db_column_tx(
 
 /// Given a table name and a new name for the table, rename the table in the database using the
 /// given database transaction.
-pub async fn rename_db_table_tx(
-    db_kind: &DbKind,
-    config: &ValveConfig,
+pub async fn rename_table_tx(
+    valve: &Valve,
     tx: &mut Transaction<'_, sqlx::Any>,
     table: &str,
-    new_tbl_name: &str,
+    new_name: &str,
 ) -> Result<()> {
+    // We begin by saving the row number and row order of the table to be renamed:
+    let (saved_rn, saved_ro) = {
+        let sql = local_sql_syntax(
+            &valve.db_kind,
+            &format!(
+                r#"SELECT "row_number", "row_order" FROM "table"
+                        WHERE "table" = {SQL_PARAM}"#
+            ),
+        );
+        let row = sqlx_query(&sql)
+            .bind(table)
+            .fetch_one(tx.acquire().await?)
+            .await?;
+        let rn: i64 = row.try_get::<i64, &str>("row_number")?;
+        let ro: i64 = row.try_get::<i64, &str>("row_order")?;
+        (rn, ro)
+    };
+
+    let cols_for_query = valve
+        .config
+        .table
+        .get("table")
+        .expect("Table table not found in config")
+        .column
+        .keys()
+        .filter(|col| *col != "table")
+        .map(|col| format!(r#""{col}""#))
+        .collect::<Vec<_>>();
+
+    let main_column_list = cols_for_query.join(", ");
+    let t2_column_list = cols_for_query
+        .iter()
+        .map(|col| format!("t2.{col}"))
+        .collect::<Vec<_>>()
+        .join(", ");
+
+    let (rn, ro) = get_next_new_row_tx(&valve.db_kind, tx, "table").await?;
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"INSERT INTO "table" (
+                       "row_number", "row_order", "table",
+                       {main_column_list}
+                   )
+                   SELECT
+                      {rn} AS row_number,
+                      {ro} as row_order,
+                      {SQL_PARAM} as "table",
+                      {t2_column_list}
+                     FROM "table" t2
+                    WHERE t2."table" = {SQL_PARAM}"#
+        ),
+    );
+    let query = sqlx_query(&sql).bind(new_name).bind(table);
+    query.execute(tx.acquire().await?).await?;
+
+    // Update from() structures in the column table that refer to the column.
+    let mut structure_params = vec![];
+    let structure_sql = {
+        let mut sql_lines = vec![];
+        for (_, parsed) in valve.structure_conditions.iter() {
+            if let Expression::Function(name, args) = &parsed.parsed {
+                if name == "from" {
+                    if let Expression::Field(ftable, fcolumn) = &*args[0] {
+                        if ftable == table {
+                            sql_lines.push(format!(
+                                r#"WHEN "structure" = {SQL_PARAM} THEN {SQL_PARAM}"#
+                            ));
+                            structure_params.push(parsed.original.to_string());
+                            let new_name = {
+                                if new_name.contains(char::is_whitespace) {
+                                    &format!("'{new_name}'")
+                                } else {
+                                    new_name
+                                }
+                            };
+                            let new_cond = format!("from({new_name}.{fcolumn})");
+                            structure_params.push(new_cond);
+                        }
+                    }
+                }
+            }
+        }
+
+        // Only actually build the clause if there were matching conditions:
+        if !sql_lines.is_empty() {
+            format!(
+                r#", "structure" = CASE
+                          {case_clause}
+                          ELSE "structure"
+                        END"#,
+                case_clause = sql_lines.join("\n")
+            )
+        } else {
+            "".to_string()
+        }
+    };
+
+    // Update column table:
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "column"
+                      SET "table" = CASE
+                        WHEN "table" = {SQL_PARAM} THEN {SQL_PARAM}
+                        ELSE "table"
+                      END{structure_sql}"#
+        ),
+    );
+    let mut query = sqlx_query(&sql).bind(table).bind(new_name);
+    for param in &structure_params {
+        query = query.bind(param);
+    }
+    query.execute(tx.acquire().await?).await?;
+
+    // Update rule table:
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(
+            r#"UPDATE "rule"
+                      SET "table" = {SQL_PARAM}
+                    WHERE "table" = {SQL_PARAM}"#
+        ),
+    );
+    let query = sqlx_query(&sql).bind(new_name).bind(table);
+    query.execute(tx.acquire().await?).await?;
+
+    let sql = local_sql_syntax(
+        &valve.db_kind,
+        &format!(r#"DELETE FROM "table" WHERE "table" = {SQL_PARAM}"#),
+    );
+    let query = sqlx_query(&sql).bind(table);
+    query.execute(tx.acquire().await?).await?;
+
+    // Finally update the newly added table's row number and row order to the values
+    // that we saved earlier, to complete the logical rename operation:
+    let sql = format!(
+        r#"UPDATE "table"
+                  SET "row_number" = {saved_rn}, "row_order" = {saved_ro}
+                WHERE "row_number" = {rn}"#
+    );
+    let query = sqlx_query(&sql);
+    query.execute(tx.acquire().await?).await?;
+
     // Although PostgreSQL supports an ALTER VIEW statement, SQLite does not, so we have to
     // manually drop and recreate views in that case. For simplicitly, then, we use the manual
     // method in both cases.
@@ -2165,16 +2683,22 @@ pub async fn rename_db_table_tx(
 
     // Next, rename the table and its corresponding conflict table:
     for suffix in ["", "_conflict"] {
-        let sql = format!(r#"ALTER TABLE "{table}{suffix}" RENAME TO "{new_tbl_name}{suffix}""#);
+        let sql = format!(r#"ALTER TABLE "{table}{suffix}" RENAME TO "{new_name}{suffix}""#);
         let query = sqlx_query(&sql);
         query.execute(tx.acquire().await?).await?;
     }
 
     // Now recreate both views:
-    let sql = get_sql_for_standard_view(new_tbl_name, db_kind);
+    let sql = get_sql_for_standard_view(new_name, &valve.db_kind);
     let query = sqlx_query(&sql);
     query.execute(tx.acquire().await?).await?;
-    let sql = get_sql_for_text_view(&config.table, table, Some(new_tbl_name), None, db_kind);
+    let sql = get_sql_for_text_view(
+        &valve.config.table,
+        table,
+        Some(new_name),
+        None,
+        &valve.db_kind,
+    );
     let query = sqlx_query(&sql);
     query.execute(tx.acquire().await?).await?;
 
@@ -3655,6 +4179,165 @@ pub async fn update_row_tx(
         false,
     )
     .await?;
+
+    Ok(())
+}
+
+/// TODO: Add docstring
+pub async fn add_column_tx(
+    table: &str,
+    column: &str,
+    column_details: &SerdeMap,
+    tx: &mut Transaction<'_, sqlx::Any>,
+    db_kind: &DbKind,
+) -> Result<u32> {
+    let make_err = |err_str: &str| -> ValveError { ValveError::InputError(err_str.to_string()) };
+
+    let datatype = match column_details.get("datatype") {
+        Some(datatype) => datatype.as_str().ok_or(make_err("Not a string"))?,
+        None => return Err(make_err("No datatype given").into()),
+    };
+    let label = column_details.get("label").and_then(|l| l.as_str());
+    let nulltype = column_details.get("nulltype").and_then(|l| l.as_str());
+    let structure = column_details.get("structure").and_then(|l| l.as_str());
+    let description = column_details.get("description").and_then(|l| l.as_str());
+
+    // Generate an insert statement and execute it:
+    let mut fields = vec![r#""table""#, r#""column""#, r#""datatype""#];
+    let mut placeholders = vec![SQL_PARAM, SQL_PARAM, SQL_PARAM];
+    let mut field_params = vec![table, column, datatype];
+    for (field, field_param) in [
+        (r#""label""#, label),
+        (r#""nulltype""#, nulltype),
+        (r#""structure""#, structure),
+        (r#""description""#, description),
+    ] {
+        if let Some(param) = field_param {
+            fields.push(field);
+            if param == "" {
+                placeholders.push("NULL");
+            } else {
+                placeholders.push(SQL_PARAM);
+                field_params.push(param);
+            }
+        }
+    }
+    let (rn, ro) = get_next_new_row_tx(db_kind, tx, "column").await?;
+    let sql = local_sql_syntax(
+        db_kind,
+        &format!(
+            r#"INSERT INTO "column" ("row_number", "row_order", {fields})
+                   VALUES ({rn}, {ro}, {placeholders})"#,
+            fields = fields.join(", "),
+            placeholders = placeholders.join(", ")
+        ),
+    );
+    let mut query = sqlx_query(&sql);
+    for param in &field_params {
+        query = query.bind(param);
+    }
+    query.execute(tx.acquire().await?).await?;
+
+    Ok(rn)
+}
+
+/// TODO: Add docstring
+pub async fn add_datatype_tx(
+    dt_map: &HashMap<String, String>,
+    db_kind: &DbKind,
+    tx: &mut Transaction<'_, sqlx::Any>,
+) -> Result<()> {
+    // Separate the datatype fields into the columns and column parameters that we will
+    // use to construct the INSERT statement:
+    let mut columns = vec![];
+    let mut placeholders = vec![];
+    let mut params = vec![];
+    for (column, value) in dt_map.iter() {
+        columns.push(format!(r#""{column}""#));
+        if value == "" {
+            placeholders.push("NULL");
+        } else {
+            placeholders.push(SQL_PARAM);
+            params.push(value);
+        }
+    }
+
+    // Generate an INSERT statement:
+    let (rn, ro) = get_next_new_row_tx(db_kind, tx, "datatype").await?;
+    let sql = local_sql_syntax(
+        db_kind,
+        &format!(
+            r#"INSERT INTO "datatype" ("row_number", "row_order", {columns})
+               VALUES ({rn}, {ro}, {placeholders})"#,
+            columns = columns.join(", "),
+            placeholders = placeholders.join(", "),
+        ),
+    );
+
+    // Insert the datatype:
+    let mut query = sqlx_query(&sql);
+    for param in params {
+        query = query.bind(param);
+    }
+    query.execute(tx.acquire().await?).await?;
+
+    Ok(())
+}
+
+/// TODO: Add docstring
+pub async fn delete_column_tx(
+    table: &str,
+    column: &str,
+    tx: &mut Transaction<'_, sqlx::Any>,
+    db_kind: &DbKind,
+) -> Result<()> {
+    let sql = local_sql_syntax(
+        db_kind,
+        &format!(r#"DELETE FROM "column" WHERE "table" = {SQL_PARAM} AND "column" = {SQL_PARAM}"#),
+    );
+
+    let query = sqlx_query(&sql).bind(table).bind(column);
+    query.execute(tx.acquire().await?).await?;
+
+    Ok(())
+}
+
+/// TODO: Add docstring
+pub async fn delete_datatype_tx(
+    datatype: &str,
+    tx: &mut Transaction<'_, sqlx::Any>,
+    db_kind: &DbKind,
+) -> Result<()> {
+    let sql = local_sql_syntax(
+        db_kind,
+        &format!(r#"DELETE FROM "datatype" WHERE "datatype" = {SQL_PARAM}"#),
+    );
+    let query = sqlx_query(&sql).bind(datatype);
+    query.execute(tx.acquire().await?).await?;
+
+    Ok(())
+}
+
+pub async fn delete_table_tx(
+    table: &str,
+    tx: &mut Transaction<'_, sqlx::Any>,
+    db_kind: &DbKind,
+) -> Result<()> {
+    // Delete it from the column table:
+    let sql = local_sql_syntax(
+        db_kind,
+        &format!(r#"DELETE FROM "column" WHERE "table" = {SQL_PARAM}"#),
+    );
+    let query = sqlx_query(&sql).bind(table);
+    query.execute(tx.acquire().await?).await?;
+
+    // Delete it from tbe table table:
+    let sql = local_sql_syntax(
+        db_kind,
+        &format!(r#"DELETE FROM "table" WHERE "table" = {SQL_PARAM}"#),
+    );
+    let query = sqlx_query(&sql).bind(table);
+    query.execute(tx.acquire().await?).await?;
 
     Ok(())
 }
